dataset_pipeline/requirements.txt
openai
easyocr
torch
torchvision
Pillow
requests
